Добавлена валидация для конфигов User и Mongo
