+++ project/libs/shared/core/src/lib/constants/app.constant.ts
@@ -1,0 +1,15 @@
+export const MIN_PORT = 0
+export const MAX_PORT = 65535
+
+export const Environment = {
+  DEV: 'development',
+  PROD: 'production',
+  STAGE: 'stage'
+} as const;
+
+export type EnvironmentType = typeof Environment[keyof typeof Environment];
+
+export const ConfigEnvironment = {
+  USER: 'user',
+  DB: 'database'
+} as const;
