+++ project/libs/shared/data-access/src/index.ts
@@ -1,3 +1,3 @@
 export { Repository } from './lib/repository/repository.interface'
 export { BaseMemoryRepository } from './lib/repository/base-memory.repository'
-export { MongoConfig } from './lib/repository/config/mongodb/mongodb.config'
+export { default as mongoConfig } from './lib/repository/config/mongodb/mongodb.config'
