+++ project/libs/user/config/src/index.ts
@@ -1,4 +1,3 @@
 export { UserConfigModule } from './lib/user-config.module'
-export { ConfigEnvironment } from './lib/user-config.constant'
-export { default as appConfig } from './lib/configurations/app.config'
-export { default as dbConfig } from './lib/configurations/db.config'
+export { default as userConfig } from './lib/user.config'
+
