+++ project/libs/user/config/src/lib/user-config.constant.ts
@@ -1,15 +1,10 @@
+import { MessagesType } from '@project/shared/core';
+
 export const USERS_ENV_FILE_PATH = 'apps/user/user.env'
 export const DEFAULT_PORT = 8000;
-export const DEFAULT_MONGODB_PORT = 27017;
 
-export const Environment = {
-  DEV: 'development',
-  PROD: 'production',
-  STAGE: 'stage'
-} as const;
-export type EnvironmentType = typeof Environment[keyof typeof Environment];
-
-export const ConfigEnvironment = {
-  APP: 'application',
-  DB: 'database'
+export const UserConfigMessage: MessagesType = {
+  ERROR: {
+    USER_APP_HOST_REQUIRED: '[User App Config] host name is required',
+  },
 } as const;
