+++ project/libs/user/config/src/lib/user-config.module.ts
@@ -1,16 +1,15 @@
 import { Module } from '@nestjs/common'
 import { ConfigModule } from '@nestjs/config'
 import { USERS_ENV_FILE_PATH } from './user-config.constant'
-import appConfig from './configurations/app.config'
-import dbConfig from './configurations/db.config'
-
+import { mongoConfig } from '@project/shared/data-access'
+import userConfig from './user.config'
 @Module({
   imports: [
     ConfigModule.forRoot({
       isGlobal: true,
       cache: true,
       // Cписок конфигураций для загрузки
-      load: [appConfig, dbConfig],
+      load: [userConfig, mongoConfig],
 
       envFilePath: USERS_ENV_FILE_PATH
     })
