+++ project/libs/user/config/tsconfig.lib.json
@@ -7,5 +7,5 @@
     "types": ["node"]
   },
   "exclude": ["jest.config.ts", "src/**/*.spec.ts", "src/**/*.test.ts"],
-  "include": ["src/**/*.ts"]
+  "include": ["src/**/*.ts", "../../shared/data-access/src/lib/repository/config/mongodb/mongodb.config.ts"]
 }
