+++ project/tsconfig.base.json
@@ -18,12 +18,12 @@
       "@project/authentication": ["libs/user/authentication/src/index.ts"],
       "@project/blog-post": ["libs/post/blog-post/src/index.ts"],
       "@project/blog-user": ["libs/user/blog-user/src/index.ts"],
-      "@project/comment": ["libs/post/comment/src/index.ts"],
-      "@project/config": ["libs/user/config/src/index.ts"],
-      "@project/like": ["libs/post/like/src/index.ts"],
-      "@project/password": ["libs/user/password/src/index.ts"],
-      "@project/repost": ["libs/post/repost/src/index.ts"],
-      "@project/search": ["libs/post/search/src/index.ts"],
+      "@project/post/comment": ["libs/post/comment/src/index.ts"],
+      "@project/post/repost": ["libs/post/repost/src/index.ts"],
+      "@project/post/search": ["libs/post/search/src/index.ts"],
+      "@project/post/like": ["libs/post/like/src/index.ts"],
+      "@project/user/config": ["libs/user/config/src/index.ts"],
+      "@project/user/password": ["libs/user/password/src/index.ts"],
       "@project/shared/core": ["libs/shared/core/src/index.ts"],
       "@project/shared/data-access": ["libs/shared/data-access/src/index.ts"],
       "@project/shared/hasher": ["libs/shared/hasher/src/index.ts"],
